feat(ui): TE-2112 move rca what-where pages to nested folder
