+++ thirdeye-ui/src/app/pages/rca/what-where-page/dimension-analysis-drills-page/dimension-analysis-drills-page.component.tsx
@@ -17,62 +17,52 @@ import { every, isEmpty } from "lodash";
 import React, { FunctionComponent, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useOutletContext, useParams, useSearchParams } from "react-router-dom";
-import { EmptyStateSwitch } from "../../../components/page-states/empty-state-switch/empty-state-switch.component";
-import { LoadingErrorStateSwitch } from "../../../components/page-states/loading-error-state-switch/loading-error-state-switch.component";
-import { BaselineOffsetSelection } from "../../../components/rca/analysis-tabs/baseline-offset-selection/baseline-offset-selection.component";
-import { AnomalyFilterOption } from "../../../components/rca/anomaly-dimension-analysis/anomaly-dimension-analysis.interfaces";
-import { DimensionSearchAutocomplete } from "../../../components/rca/heat-map/dimension-search-automcomplete/dimension-search-autocomplete.component";
-import { HeatMap } from "../../../components/rca/heat-map/heat-map.component";
-import { formatDimensionOptions } from "../../../components/rca/heat-map/heat-map.utils";
-import { PreviewChart } from "../../../components/rca/top-contributors-table/preview-chart/preview-chart.component";
+import { EmptyStateSwitch } from "../../../../components/page-states/empty-state-switch/empty-state-switch.component";
+import { LoadingErrorStateSwitch } from "../../../../components/page-states/loading-error-state-switch/loading-error-state-switch.component";
+import { AnomalyFilterOption } from "../../../../components/rca/anomaly-dimension-analysis/anomaly-dimension-analysis.interfaces";
+import { DimensionSearchAutocomplete } from "../../../../components/rca/heat-map/dimension-search-automcomplete/dimension-search-autocomplete.component";
+import { formatDimensionOptions } from "../../../../components/rca/heat-map/heat-map.utils";
+import { PreviewChart } from "../../../../components/rca/top-contributors-table/preview-chart/preview-chart.component";
 import {
     PageContentsCardV1,
     SkeletonV1,
     useNotificationProviderV1,
-} from "../../../platform/components";
-import { ActionStatus } from "../../../rest/actions.interfaces";
+} from "../../../../platform/components";
+import { ActionStatus } from "../../../../rest/actions.interfaces";
 import {
     AnomalyBreakdown,
     SavedStateKeys,
-} from "../../../rest/dto/rca.interfaces";
-import { useGetAnomalyMetricBreakdown } from "../../../rest/rca/rca.actions";
-import { areFiltersEqual } from "../../../utils/anomaly-dimension-analysis/anomaly-dimension-analysis";
-import { getFromSavedInvestigationOrDefault } from "../../../utils/investigation/investigation.util";
-import { notifyIfErrors } from "../../../utils/notifications/notifications.util";
+} from "../../../../rest/dto/rca.interfaces";
+import { useGetAnomalyMetricBreakdown } from "../../../../rest/rca/rca.actions";
+import { areFiltersEqual } from "../../../../utils/anomaly-dimension-analysis/anomaly-dimension-analysis";
+import { getFromSavedInvestigationOrDefault } from "../../../../utils/investigation/investigation.util";
+import { notifyIfErrors } from "../../../../utils/notifications/notifications.util";
 import {
     concatKeyValueWithEqual,
     deserializeKeyValuePair,
     serializeKeyValuePair,
-} from "../../../utils/params/params.util";
-import { RootCauseAnalysisForAnomalyPageParams } from "../../root-cause-analysis-for-anomaly-page/root-cause-analysis-for-anomaly-page.interfaces";
-import { InvestigationContext } from "../investigation-state-tracker-container-page/investigation-state-tracker.interfaces";
+} from "../../../../utils/params/params.util";
+import { RootCauseAnalysisForAnomalyPageParams } from "../../../root-cause-analysis-for-anomaly-page/root-cause-analysis-for-anomaly-page.interfaces";
+import { InvestigationContext } from "../../investigation-state-tracker-container-page/investigation-state-tracker.interfaces";
 
-const HEATMAP_FILTERS_URL_KEY = "heatmapFilters";
+const QUERY_PARAM_KEY = "dimensionAnalysisFilters";
 
-export const HeatMapPage: FunctionComponent = () => {
+export const DimensionAnalysisDrillsPage: FunctionComponent = () => {
     const { notify } = useNotificationProviderV1();
     const { t } = useTranslation();
     const [searchParams, setSearchParams] = useSearchParams();
     const { id: anomalyId } =
         useParams<RootCauseAnalysisForAnomalyPageParams>();
 
-    // Keep track of last height to use as loading state height
-    const [lastTreemapHeight, setLastTreemapHeight] = useState(300);
-
     const [anomalyFilters, setAnomalyFilters] = useState<AnomalyFilterOption[]>(
         () => {
-            const heatmapFilterQueryParams = searchParams.get(
-                HEATMAP_FILTERS_URL_KEY
-            );
+            const heatmapFilterQueryParams = searchParams.get(QUERY_PARAM_KEY);
 
             return heatmapFilterQueryParams
                 ? deserializeKeyValuePair(heatmapFilterQueryParams)
                 : [];
         }
     );
-    const [comparisonOffset, setComparisonOffset] = useState(() => {
-        return searchParams.get("baselineWeekOffset") ?? "P1W";
-    });
 
     const { investigation, anomaly, alertInsight, onInvestigationChange } =
         useOutletContext<InvestigationContext>();
@@ -86,9 +76,7 @@ export const HeatMapPage: FunctionComponent = () => {
 
     // Sync the anomaly filters if the search params changed
     useEffect(() => {
-        const currentQueryFilterSearchQuery = searchParams.get(
-            HEATMAP_FILTERS_URL_KEY
-        );
+        const currentQueryFilterSearchQuery = searchParams.get(QUERY_PARAM_KEY);
 
         if (
             currentQueryFilterSearchQuery &&
@@ -108,14 +96,14 @@ export const HeatMapPage: FunctionComponent = () => {
 
     useEffect(() => {
         getMetricBreakdown(Number(anomalyId), {
-            baselineOffset: comparisonOffset,
+            baselineOffset: "P1W",
             filters: [
                 ...anomalyFilters.map((item) =>
                     concatKeyValueWithEqual(item, false)
                 ),
             ],
         });
-    }, [anomalyId, comparisonOffset, anomalyFilters]);
+    }, [anomalyId, anomalyFilters]);
 
     useEffect(() => {
         notifyIfErrors(
@@ -127,12 +115,6 @@ export const HeatMapPage: FunctionComponent = () => {
             })
         );
     }, [heatMapDataRequestStatus]);
-
-    const handleBaselineChange = (newValue: string): void => {
-        setComparisonOffset(newValue);
-        searchParams.set("baselineWeekOffset", newValue);
-        setSearchParams(searchParams);
-    };
 
     const handleAddDimensionsToInvestigationClick = (): void => {
         const currentDimensions = getFromSavedInvestigationOrDefault<
@@ -160,10 +142,10 @@ export const HeatMapPage: FunctionComponent = () => {
 
     const handleFilterChange = (newFilters: AnomalyFilterOption[]): void => {
         if (newFilters.length === 0) {
-            searchParams.delete(HEATMAP_FILTERS_URL_KEY);
+            searchParams.delete(QUERY_PARAM_KEY);
         } else {
             searchParams.set(
-                HEATMAP_FILTERS_URL_KEY,
+                QUERY_PARAM_KEY,
                 serializeKeyValuePair(newFilters)
             );
         }
@@ -174,11 +156,11 @@ export const HeatMapPage: FunctionComponent = () => {
     return (
         <>
             <Grid item xs={12}>
-                <Typography variant="h4">{t("label.heatmap")}</Typography>
+                <Typography variant="h4">
+                    {t("label.dimension-analysis-drills")}
+                </Typography>
                 <Typography variant="body1">
-                    {t(
-                        "message.represents-the-frequency-and-severity-of-a-change"
-                    )}
+                    {t("message.manual-filter-for-dimensions-analysis")}
                 </Typography>
             </Grid>
             <Grid item xs={12}>
@@ -194,17 +176,6 @@ export const HeatMapPage: FunctionComponent = () => {
                                     {t(
                                         "message.select-the-dimensions-below-to-drill-down-into-the"
                                     )}
-                                </Grid>
-                                <Grid item xs>
-                                    <BaselineOffsetSelection
-                                        baselineOffset={comparisonOffset}
-                                        label={t(
-                                            "label.dimensions-changed-from-the-last"
-                                        )}
-                                        onBaselineOffsetChange={
-                                            handleBaselineChange
-                                        }
-                                    />
                                 </Grid>
                             </Grid>
                         </Grid>
@@ -230,7 +201,7 @@ export const HeatMapPage: FunctionComponent = () => {
                                     <Box pb={2} pt={2}>
                                         <SkeletonV1
                                             animation="pulse"
-                                            height={lastTreemapHeight}
+                                            height={300}
                                             variant="rect"
                                         />
                                     </Box>
@@ -258,39 +229,32 @@ export const HeatMapPage: FunctionComponent = () => {
                                         )
                                     }
                                 >
-                                    <HeatMap
-                                        anomalyFilters={anomalyFilters}
-                                        heatMapData={
-                                            heatMapData as AnomalyBreakdown
-                                        }
-                                        onFilterChange={handleFilterChange}
-                                        onHeightChange={setLastTreemapHeight}
-                                    />
-                                </EmptyStateSwitch>
-
-                                <Box pt={2}>
-                                    <PreviewChart
-                                        alertInsight={alertInsight}
-                                        anomaly={anomaly}
-                                        dimensionCombinations={
-                                            isEmpty(anomalyFilters)
-                                                ? []
-                                                : [anomalyFilters]
-                                        }
-                                    >
-                                        <Button
-                                            color="primary"
-                                            disabled={isEmpty([anomalyFilters])}
-                                            onClick={
-                                                handleAddDimensionsToInvestigationClick
+                                    <Box pt={2}>
+                                        <PreviewChart
+                                            alertInsight={alertInsight}
+                                            anomaly={anomaly}
+                                            dimensionCombinations={
+                                                isEmpty(anomalyFilters)
+                                                    ? []
+                                                    : [anomalyFilters]
                                             }
                                         >
-                                            {t(
-                                                "label.add-dimensions-to-investigation"
-                                            )}
-                                        </Button>
-                                    </PreviewChart>
-                                </Box>
+                                            <Button
+                                                color="primary"
+                                                disabled={isEmpty([
+                                                    anomalyFilters,
+                                                ])}
+                                                onClick={
+                                                    handleAddDimensionsToInvestigationClick
+                                                }
+                                            >
+                                                {t(
+                                                    "label.add-dimensions-to-investigation"
+                                                )}
+                                            </Button>
+                                        </PreviewChart>
+                                    </Box>
+                                </EmptyStateSwitch>
                             </LoadingErrorStateSwitch>
                         </Grid>
                     </Grid>
